backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-docx
pydantic>=2.9.0
python-multipart
python-dotenv
